Api/requirements.txt
cassandra-driver 
fastapi 
uvicorn